Detect -1 properly (it maps to -0.01)
git-svn-id: file:///home/svn/DVA2006@18424 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/taglib/format/LandingScoreFormatTag.java
@@ -66,7 +66,7 @@ public class LandingScoreFormatTag extends NumberFormatTag {
 	
 	@Override
 	public int doEndTag() throws JspException {
-		boolean noValue = (_value.intValue() < 0);
+		boolean noValue = (_value.doubleValue() < 0);
 		if (noValue && StringUtils.isEmpty(_default))
 			return EVAL_PAGE;
